requirements.txt
Flask==0.10.1
Flask-Moment==0.5.0
Flask-RESTful==0.3.3
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
Flask-WhooshAlchemy==0.56
Jinja2==2.7.1
SQLAlchemy==0.9.9
sqlalchemy-migrate
pandas==0.16.2
psycopg2==2.6.1
tornado==4.2.1
requests==2.1.0
bson
pymongo
gunicorn
